• when posting to WP: add a newline before the more part.
The reason is that WP will add <a id="more"></a> between the main entry and the main entry continued. If one uses Markdown this will result in the line becoming <p><a id="more"></a><br /> [1] but with the extra newline, it instead becomes <p><a id="more"></a></p> which doesn’t result in extra space (in WebKit.)

Ideally the Markdown plug-in should probably be changed, but I don’t think there are any bad side effects of this extra leading newline.

[1] this behavior though only if you’ve patched the markdown plug-in to provide the priority (30) when it attempts to remove_filter the wpautop function.


git-svn-id: http://svn.textmate.org/trunk/Bundles/Blogging.tmbundle@4014 dfb7d73b-c2ec-0310-8fea-fb051d288c6d